refactor: Move logic to Home page and add loading
